requirements.txt
boto3==1.34.88
codetiming==1.4.0
cryptography==42.0.5
Django==4.2.11
dj-database-url==2.1.0
django-allauth==0.61.1
django-cors-headers==4.3.1
django-csp==3.8
django-debug-toolbar==4.3.0
django-filter==24.2
django-ipware==6.0.5
python-ipware==2.0.3  # 2.0.4, 2.0.5 break tests
django-redis==5.4.0
django-ftl==0.14
django-referrer-policy==1.0
djangorestframework==3.15.1
django-waffle==4.1.0
dockerflow==2024.4.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
glean_parser==14.0.1
google-measurement-protocol==1.1.0
google-cloud-profiler==4.1.0
gunicorn==22.0.0
jwcrypto==1.5.6
markus[datadog]==4.2.0
pem==23.1.0
psycopg[c]==3.1.18
PyJWT==2.8.0
python-decouple==3.8
pyOpenSSL==24.1.0
requests==2.31.0
sentry-sdk==1.45.0
whitenoise==6.6.0

# phones app
phonenumbers==8.13.35
twilio==9.0.5
vobject==0.9.7

# tests
coverage==7.4.4
model-bakery==1.18.0
pytest-cov==5.0.0
pytest-django==4.8.0
responses==0.25.0

# linting
black==24.4.0
ruff==0.3.6

# type hinting
boto3-stubs==1.34.88
botocore-stubs==1.34.88
django-stubs==4.2.6
djangorestframework-stubs==3.14.4
mypy-boto3-ses==1.34.0
mypy==1.9.0
types-pyOpenSSL==24.0.0.20240417
types-requests==2.31.0.20240406
